quicksafe

+++ baseline/device.cuh
@@ -2,6 +2,6 @@
 #define H_DEVICE _
 #include "../shared/shared.cuh"
 
-__global__ void scan_lookback(float *a, PartitionDescriptor *states);
+__global__ void scan_lookback(int *a, PartitionDescriptor *states);
 
 #endif
